+++ farmbot_core/priv/config/migrations/20180202171932_add_network_interface_deprication.exs
@@ -1,0 +1,9 @@
+defmodule Farmbot.Config.Repo.Migrations.AddNetworkInterfaceDeprication do
+  use Ecto.Migration
+
+  def change do
+    alter table("network_interfaces") do
+      add(:migrated, :boolean)
+    end
+  end
+end