+++ farmbot_core/priv/config/migrations/20180403180213_add_sync_timeout.exs
@@ -1,0 +1,8 @@
+defmodule Farmbot.Config.Repo.Migrations.AddSyncTimeout do
+  use Ecto.Migration
+  import Farmbot.Config.MigrationHelpers
+
+  def change do
+    create_settings_config("sync_timeout_ms", :float, 90_000.00)
+  end
+end