+++ farmbot_core/priv/config/migrations/20180627122653_add_secret_field.exs
@@ -1,0 +1,8 @@
+defmodule Farmbot.Config.Repo.Migrations.AddSecretField do
+  use Ecto.Migration
+  import Farmbot.Config.MigrationHelpers
+
+  def change do
+    create_auth_config("secret", :string, nil)
+  end
+end