+++ farmbot_os/platform/host/system_tasks.ex
@@ -1,0 +1,13 @@
+defmodule Farmbot.Host.SystemTasks do
+  @moduledoc "Host implementation for Farmbot.System."
+
+  @behaviour Farmbot.System
+
+  def reboot() do
+    shutdown()
+  end
+
+  def shutdown() do
+    System.halt()
+  end
+end