+++ farmbot_os/platform/target/system_tasks.ex
@@ -1,0 +1,13 @@
+defmodule Farmbot.Target.SystemTasks do
+  @moduledoc "Target implementation for System Tasks."
+
+  @behaviour Farmbot.System
+
+  def reboot() do
+    Nerves.Runtime.reboot()
+  end
+
+  def shutdown() do
+    Nerves.Runtime.poweroff()
+  end
+end