+++ lib/farmbot/repo/module_type/farm_event.ex
@@ -1,4 +1,0 @@
-defmodule Farmbot.Repo.ModuleType.FarmEvent do
-  @moduledoc false
-  use Farmbot.Repo.ModuleType, valid_mods: ~w(Sequence Regimen)
-end