+++ test/farmbot/celery_script/ast/arg/location_test.exs
@@ -1,7 +1,0 @@
-defmodule Farmbot.CeleryScript.AST.Arg.LocationTest do
-  use FarmbotTestSupport.AST.ArgTestCase
-  alias Farmbot.CeleryScript.AST
-  alias AST.Arg.Location
-  
-  arg_is_ast(Location)
-end